Tighten up definition of and use of links
Stacks are never pushed as themselves, only as links, so remove the “code”
type from SAM.md.

PC and PC0 do not point directly to stacks, but inside them. So, do not
store them on the stack as links, but as integers.

Make POP_LINK check that it actually points to a stack, and fold
sam_find_stack into this macro, its only caller. This means that the
implementation of `DO` and `IF` can use `POP_LINK` directly.

Incidentally, _POP_INSN does not need its `err` argument any more, as it is
always `SAM_ERROR_WRONG_TYPE`.