Allocate stacks individually (“nested stacks”)
This means that each stack is one stack item. The motivation is to allow
stacks to be created and manipulated easily at run-time, which we’ll used
for straightforward implementation of more advanced SAL features, such as
closures. Remove encoding of BRA/KET instructions; stacks are just pointed
to with REF now; however, for now, keep the BRA/KET encoding in the
sam_stack_t struct so that it’s easy to see that a REF points to a stack
rather than some other instruction (that is, a REF points to something
instruction-like, always). “KET” instructions are still inferred at the
end of a stack; “BRA” no longer exists in any form.

Use sam_frame_t for DO/RET. Make “GO” do a tail-call, not a branch. Re-add
IGET and ISET instructions.